♻️ Rename debug to objects
